moving bls keystore to separate module
