+++ data/keystore/src/main/java/tech/pegasys/artemis/bls/keystore/Cipher.java
@@ -11,36 +11,36 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.util.bls.keystore;
+package tech.pegasys.artemis.bls.keystore;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.base.MoreObjects;
 import org.apache.tuweni.bytes.Bytes;
 
-public class Checksum {
-  private final ChecksumFunction checksumFunction;
-  private final Param param;
+public class Cipher {
+  private final CipherFunction cipherFunction;
+  private final CipherParam cipherParam;
   private final Bytes message;
 
   @JsonCreator
-  public Checksum(
-      @JsonProperty(value = "function", required = true) final ChecksumFunction checksumFunction,
-      @JsonProperty(value = "params", required = true) final Param param,
+  public Cipher(
+      @JsonProperty(value = "function", required = true) final CipherFunction cipherFunction,
+      @JsonProperty(value = "params", required = true) final CipherParam cipherParam,
       @JsonProperty(value = "message", required = true) final Bytes message) {
-    this.checksumFunction = checksumFunction;
-    this.param = param;
+    this.cipherFunction = cipherFunction;
+    this.cipherParam = cipherParam;
     this.message = message;
   }
 
   @JsonProperty(value = "function")
-  public ChecksumFunction getChecksumFunction() {
-    return checksumFunction;
+  public CipherFunction getCipherFunction() {
+    return cipherFunction;
   }
 
   @JsonProperty(value = "params")
-  public Param getParam() {
-    return param;
+  public CipherParam getCipherParam() {
+    return cipherParam;
   }
 
   @JsonProperty(value = "message")
@@ -51,8 +51,8 @@ public class Checksum {
   @Override
   public String toString() {
     return MoreObjects.toStringHelper(this)
-        .add("cryptoFunction", checksumFunction)
-        .add("param", param)
+        .add("function", cipherFunction)
+        .add("params", cipherParam)
         .add("message", message)
         .toString();
   }
