+++ data/keystore/src/main/java/tech/pegasys/artemis/bls/keystore/CryptoFunction.java
@@ -11,36 +11,22 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.util.bls.keystore;
+package tech.pegasys.artemis.bls.keystore;
 
 import com.fasterxml.jackson.annotation.JsonValue;
-import org.bouncycastle.crypto.Digest;
-import org.bouncycastle.crypto.util.DigestFactory;
 
-public enum Pbkdf2PseudoRandomFunction {
-  HMAC_SHA1("hmac-sha1"),
-  HMAC_SHA256("hmac-sha256"),
-  HMAC_SHA512("hmac-sha512");
+public enum CryptoFunction {
+  PBKDF2("pbkdf2"),
+  SCRYPT("scrypt");
 
   private final String jsonValue;
 
-  Pbkdf2PseudoRandomFunction(final String jsonValue) {
+  CryptoFunction(final String jsonValue) {
     this.jsonValue = jsonValue;
   }
 
   @JsonValue
   public String getJsonValue() {
     return this.jsonValue;
-  }
-
-  public Digest getDigest() {
-    switch (this) {
-      case HMAC_SHA256:
-        return DigestFactory.createSHA256();
-      case HMAC_SHA512:
-        return DigestFactory.createSHA512();
-      default:
-        return DigestFactory.createSHA1();
-    }
   }
 }
