+++ data/keystore/src/main/java/tech/pegasys/artemis/bls/keystore/KeyStore.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.util.bls.keystore;
+package tech.pegasys.artemis.bls.keystore;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
