+++ data/keystore/src/main/java/tech/pegasys/artemis/bls/keystore/KeyStoreFactory.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.util.bls.keystore;
+package tech.pegasys.artemis.bls.keystore;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.File;
