+++ settings.gradle
@@ -7,6 +7,7 @@ include 'data:metrics'
 include 'data:provider'
 include 'data:beaconrestapi'
 include 'data:recorder'
+include 'data:keystore'
 include 'events'
 include 'logging'
 include 'networking:eth2'
